feat: Add minimal sqlite module
tweak test settings

Issue #21

+++ tests/database/sqlite.py
@@ -30,7 +30,7 @@ import bbutil.lang.parser
 import bbutil.lang.parser.pyfile
 
 from unittest.mock import MagicMock, Mock
-from bbutil.database.sqlite import SQLite
+from bbutil.database import SQLite, Table, Types
 from bbutil.utils import full_path
 
 from bbutil.logging import Logging
@@ -59,12 +59,21 @@ class TestSQLite(unittest.TestCase):
         _log.open()
         return _log
 
-    def _get_sqlite(self) -> SQLite:
+    @staticmethod
+    def _get_table_01(sqlite_object: SQLite):
+        _table = Table(name="tester01", sqlite=sqlite_object)
+        _table.add_column(name="testid", data_type=Types.integer, unique=True)
+        _table.add_column(name="use_test", data_type=Types.bool)
+        _table.add_column(name="testname", data_type=Types.string)
+        _table.add_column(name="path", data_type=Types.string)
+        return
+
+    def _get_sqlite(self, filename: str, path: str = os.getcwd(), clean: bool = False) -> SQLite:
         _log = self.set_log()
-        _testfile = full_path("{0:s}/test.sqlite".format(os.getcwd()))
+        _testfile = full_path("{0:s}/{1:s}".format(path, filename))
         _name = "Test"
 
-        if os.path.exists(_testfile) is True:
+        if (os.path.exists(_testfile) is True) and (clean is True):
             os.remove(_testfile)
 
         _sqlite = SQLite(filename=_testfile, name="Test", log=_log)
@@ -165,7 +174,7 @@ class TestSQLite(unittest.TestCase):
         return
 
     def test_disconnect_01(self):
-        _sqlite = self._get_sqlite()
+        _sqlite = self._get_sqlite(filename="test.sqlite")
 
         _check1 = _sqlite.connect()
         _check2 = _sqlite.disconnect()
@@ -175,7 +184,7 @@ class TestSQLite(unittest.TestCase):
         return
 
     def test_disconnect_02(self):
-        _sqlite = self._get_sqlite()
+        _sqlite = self._get_sqlite(filename="test.sqlite")
 
         _check1 = _sqlite.connect()
         _check2 = _sqlite.disconnect()
@@ -187,7 +196,7 @@ class TestSQLite(unittest.TestCase):
         return
 
     def test_disconnect_03(self):
-        _sqlite = self._get_sqlite()
+        _sqlite = self._get_sqlite(filename="test.sqlite")
 
         _check1 = _sqlite.connect()
 
@@ -200,7 +209,7 @@ class TestSQLite(unittest.TestCase):
         return
 
     def test_disconnect_04(self):
-        _sqlite = self._get_sqlite()
+        _sqlite = self._get_sqlite(filename="test.sqlite")
 
         _check1 = _sqlite.connect()
 
@@ -217,7 +226,7 @@ class TestSQLite(unittest.TestCase):
         return
 
     def test_disconnect_05(self):
-        _sqlite = self._get_sqlite()
+        _sqlite = self._get_sqlite(filename="test.sqlite")
 
         _check1 = _sqlite.connect()
 
@@ -231,4 +240,17 @@ class TestSQLite(unittest.TestCase):
 
         self.assertTrue(_check1)
         self.assertFalse(_check2)
+        return
+
+    def test_check_table_01(self):
+        _sqlite = self._get_sqlite(filename="test.sqlite")
+
+        _check1 = _sqlite.connect()
+        _check2 = _sqlite.check_table("Test")
+
+        _check3 = _sqlite.disconnect()
+
+        self.assertTrue(_check1)
+        self.assertFalse(_check2)
+        self.assertTrue(_check3)
         return
